requirements.txt
# FastAPI and server
fastapi
uvicorn
python-multipart==0.0.20
jinja2==3.1.4
pydantic==2.8.2  # FIXED: Downgraded for LangChain compatibility
mangum

# LangChain and RAG - Compatible versions
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langchain-google-genai==1.0.10

# Document processing
pymupdf

# Vector store and embeddings
chromadb
sentence-transformers

# Utilities
python-dotenv==1.0.1